ASoC: TDF8532: Fix for double memory free in case of error condition of tdf8523_single_read
tdf8523_single_read is defind to return uint8_t while the return statement,
returns a int type, so in case of error checks are not followed (E.g -121 is returned as 135)
Which lead to double free of memory as its already freed in tdf8532_single_read in case of
failure.

+++ sound/soc/codecs/tdf8532.c
@@ -107,11 +107,11 @@ out:
 	return ret;
 }
 
-static uint8_t tdf8532_single_read(struct tdf8532_priv *dev_data,
+static int tdf8532_single_read(struct tdf8532_priv *dev_data,
 					char **repl_buff)
 {
 	struct device *dev = &(dev_data->i2c->dev);
-	uint8_t recv_len;
+	int recv_len;
 	int ret;
 
 	ret = tdf8532_read_wait_ack(dev_data, msecs_to_jiffies(ACK_TIMEOUT));
